Working on UTF8 Strings, Further Emit Changes

+++ quickrun.sh
@@ -1,19 +1,23 @@
 #!/bin/sh
 
-if [[ "$2" = "" ]]; then
-    ./build/debug/ilish -e "$1" > tmp.s
-else 
-    ./build/debug/ilish -f "$1" > tmp.s
+if [[ "$1" = "-f" ]]; then
+    ./build/debug/ilish -f "$2" > tmp.s
+else
+    ./build/debug/ilish -e "$2" > tmp.s
 fi
 
 cat tmp.s
 
 if [[ "$(head -n 1 tmp.s)" = ".data" ]]; then
-    gcc -pg -Og -ggdb3 tmp.s build/runtime/runtime.o -o tmp # using cc to call gas and ld
+    gcc -pg -Og -ggdb3 tmp.s runtime/runtime.c -o tmp 
     # as tmp.s -o tmp.o # using gas and ld directly
     # ld tmp.o build/runtime/runtime.o -o tmp -lc
     ./tmp
-    rm tmp &>/dev/null
+    if [[ "$1" != "-t" ]]; then
+        rm tmp &>/dev/null
+    fi
 fi
 
-rm tmp.s
+if [[ "$1" != "-t" ]]; then
+    rm tmp.s
+fi
